+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Category filter from /algos/ route (#1199)
+
 ## [0.23.1] 2022-07-26
 
 ### Changed